+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Weblog of stuff
 
+- Replace breadcrumbs with vault listings selector on vault detail pages (2026-02-11)
 - Added Twitter card and Telegram link previews to all vault pages (2026-02-10)